requirements.txt
python-telegram-bot>=21.5
Flask>=3.0.3
Telethon>=1.36.0
requests>=2.32.3
PySocks>=1.7.1
